requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
MarkupSafe==3.0.2
python-dotenv==1.0.1
typing_extensions==4.12.2
